Introduce Expectations interfaces to make clear the similarities and distinctions between the different classes
