+++ src/main/kotlin/org/vaccineimpact/api/models/AppliedExpectation.kt
@@ -1,0 +1,10 @@
+package org.vaccineimpact.api.models
+
+interface AppliedExpectation
+{
+    val expectation: Expectations
+    val applicableScenarios: List<String>
+    val disease: String
+}
+
+
