+++ src/main/kotlin/org/vaccineimpact/api/models/OutcomeExpectations.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.models
+
+data class OutcomeExpectations(override val id: Int,
+                               override val description: String,
+                               override val years: IntRange,
+                               override val ages: IntRange,
+                               override val cohorts: CohortRestriction,
+                               override val outcomes: List<String>): Expectations
