+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilityDetails.kt
@@ -1,10 +1,10 @@
 package org.vaccineimpact.api.models.responsibilities
 
-import org.vaccineimpact.api.models.Expectations
+import org.vaccineimpact.api.models.CountryOutcomeExpectations
 import org.vaccineimpact.api.models.TouchstoneVersion
 
 data class ResponsibilityDetails(
         val responsibility: Responsibility,
         val touchstoneVersion: TouchstoneVersion,
-        val expectations: Expectations
+        val expectations: CountryOutcomeExpectations
 )
